added testing custom mnist images
