--- tutorials/01-basics/feedforward_neural_network/temp-main_gpu.py
@@ -1,96 +1,0 @@
-import torch
-import torch.nn as nn
-import torchvision.datasets as dsets
-import torchvision.transforms as transforms
-from torch.autograd import Variable
-import numpy as np
-import matplotlib.pyplot as plt
-
-
-# Hyper Parameters 
-input_size = 784
-hidden_size = 500
-num_classes = 10
-num_epochs = 5
-batch_size = 100
-learning_rate = 0.001
-
-# MNIST Dataset 
-train_dataset = dsets.MNIST(root='./data', 
-                            train=True, 
-                            transform=transforms.ToTensor(),  
-                            download=True)
-
-a = np.array(train_dataset)
-a.numpy()
-plt.imshow(float(a[222][0]))
-print a[222][1]
-
-test_dataset = dsets.MNIST(root='./data', 
-                           train=False, 
-                           transform=transforms.ToTensor())
-
-# Data Loader (Input Pipeline)
-train_loader = torch.utils.data.DataLoader(dataset=train_dataset, 
-                                           batch_size=batch_size, 
-                                           shuffle=True)
-
-test_loader = torch.utils.data.DataLoader(dataset=test_dataset, 
-                                          batch_size=batch_size, 
-                                          shuffle=False)
-
-# Neural Network Model (1 hidden layer)
-class Net(nn.Module):
-    def __init__(self, input_size, hidden_size, num_classes):
-        super(Net, self).__init__()
-        self.fc1 = nn.Linear(input_size, hidden_size) 
-        self.relu = nn.ReLU()
-        self.fc2 = nn.Linear(hidden_size, num_classes)  
-    
-    def forward(self, x):
-        out = self.fc1(x)
-        out = self.relu(out)
-        out = self.fc2(out)
-        return out
-    
-net = Net(input_size, hidden_size, num_classes)
-net.cuda()   
-    
-# Loss and Optimizer
-criterion = nn.CrossEntropyLoss()  
-optimizer = torch.optim.Adam(net.parameters(), lr=learning_rate)  
-
-# Train the Model
-
-
-for epoch in range(num_epochs):
-    for i, (images, labels) in enumerate(train_loader):  
-        # Convert torch tensor to Variable
-        images = Variable(images.view(-1, 28*28).cuda())
-        labels = Variable(labels.cuda())
-        
-        # Forward + Backward + Optimize
-        optimizer.zero_grad()  # zero the gradient buffer
-        outputs = net(images)
-        loss = criterion(outputs, labels)
-        loss.backward()
-        optimizer.step()
-        
-        if (i+1) % 100 == 0:
-            print ('Epoch [%d/%d], Step [%d/%d], Loss: %.4f' 
-                   %(epoch+1, num_epochs, i+1, len(train_dataset)//batch_size, loss.data[0]))
-
-# Test the Model
-correct = 0
-total = 0
-for images, labels in test_loader:
-    images = Variable(images.view(-1, 28*28)).cuda()
-    outputs = net(images)
-    _, predicted = torch.max(outputs.data, 1)
-    total += labels.size(0)
-    correct += (predicted.cpu() == labels).sum()
-
-print('Accuracy of the network on the 10000 test images: %d %%' % (100 * correct / total))
-
-# Save the Model
-torch.save(net.state_dict(), 'model.pkl')
